requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
aiohttp==3.13.5
google-genai==1.73.1
pydantic==2.12.5
python-dotenv==1.2.2
httpx==0.28.1
pytest==9.0.3
pytest-asyncio==1.3.0
